feat: paleta dinamica de cores refatoradas
