+++ apps/medicamentar-desktop/src/constants/theme/ThemeContext.tsx
@@ -1,34 +1,10 @@
-import React, { createContext, useState, useMemo, useContext } from 'react';
-import { ThemeProvider as MuiThemeProvider, createTheme, Theme } from '@mui/material/styles';
-import { lightTheme, darkTheme } from './themeConfig';
+import { createContext } from "react";
 
 interface ThemeContextType {
   darkMode: boolean;
   toggleTheme: () => void;
 }
 
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (!context) throw new Error('useTheme must be used within a ThemeProvider');
-  return context;
-};
-
-export const ThemeProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-  const [darkMode, setDarkMode] = useState(false);
-
-  const theme = useMemo<Theme>(() => {
-    return createTheme(darkMode ? darkTheme : lightTheme);
-  }, [darkMode]);
-
-  const toggleTheme = () => setDarkMode(prev => !prev);
-
-  return (
-    <ThemeContext.Provider value={{ darkMode, toggleTheme }}>
-      <MuiThemeProvider theme={theme}>
-        {children}
-      </MuiThemeProvider>
-    </ThemeContext.Provider>
-  );
-};
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
