+++ apps/medicamentar-desktop/src/main.tsx
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom/client";
 import Router from "./pages/router.tsx";
 import { CssBaseline } from "@mui/material";
 import { BrowserRouter } from "react-router-dom";
-import { ThemeProvider } from "./constants/theme/ThemeContext.tsx";
+import { ThemeProvider } from "./constants/theme/ThemeProvider.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
